+++ src/main/java/DataCollection/api/DataCollectionApiClient.java
@@ -1,9 +1,6 @@
 package DataCollection.api;
 
-import DataCollection.domain.LeagueEntryDto;
-import DataCollection.domain.Match;
-import DataCollection.domain.MatchDetail;
-import DataCollection.domain.Summoners;
+import DataCollection.domain.*;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -26,6 +23,7 @@ public class DataCollectionApiClient {
     private String GET_USERNAME_URI="https://kr.api.riotgames.com/lol/league-exp/v4/entries/RANKED_SOLO_5x5/{tier}/{division}?page={page}&api_key={ApiKey}";
     private String GET_GAMEIDS_URI="https://kr.api.riotgames.com/lol/match/v4/matchlists/by-account/{AccountId}?queue=420&api_key={ApiKey}";
     private String GET_ACCOUNTID_URI="https://kr.api.riotgames.com/lol/summoner/v4/summoners/by-name/{SummonerName}?api_key={ApiKey}";
+    private String GET_TIMELINE_URI="https://kr.api.riotgames.com/lol/match/v4/timelines/by-match/{matchid}?api_key={ApiKey}";
 
     public MatchDetail getMatchDetail(long matchid){
         MatchDetail matchDetail = restTemplate.getForObject(GET_GAMEDETAIL_URI,MatchDetail.class,matchid,ApiKey);
@@ -47,6 +45,11 @@ public class DataCollectionApiClient {
                 .map(match_ -> match_.getGameId())
                 .collect(Collectors.toList());
         return matchIds;
+    }
+
+    public TimeLine getTimeLine(long matchid){
+        TimeLine timeLine = restTemplate.getForObject(GET_TIMELINE_URI,TimeLine.class,matchid,ApiKey);
+        return timeLine;
     }
 
 }
